// fe/src/lib/api/generated/model/reviewStatusValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type ReviewStatusValue = typeof ReviewStatusValue[keyof typeof ReviewStatusValue];


export const ReviewStatusValue = {
  OPEN: 'OPEN',
  RESOLVED: 'RESOLVED',
} as const;